@@ -1,5 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+
 import './index.css';
 // import Lesson1 from './components/Lesson1';
 import reportWebVitals from './reportWebVitals';
@@ -7,14 +8,13 @@ import reportWebVitals from './reportWebVitals';
 // import Airbnb from './components/airbnb/Airbnb';
 import TravelJournal from './components/TravelJournal/TravelJournal';
 
-ReactDOM.render(
-  <React.StrictMode>
-    {/* <Lesson1 /> */}
-    {/* <ReactFacts24 /> */}
-    {/* <Airbnb /> */}
-    <TravelJournal />
-  </React.StrictMode>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
+  // <Lesson1 /> */}
+  // <ReactFacts24 /> */}
+  // <Airbnb /> */}
+  <TravelJournal />
 );
 
 // If you want to start measuring performance in your app, pass a function
